services/search-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-search-results==2.4.2
python-dotenv==1.0.0
pydantic==2.6.1
requests==2.31.0
